add reference samples for testing and add exact match filtering in separate class
