+++ tests/data/reference_sample.fna
@@ -1,0 +1,44 @@
+>NM_130786.4 Homo sapiens alpha-1-B glycoprotein (A1BG), mRNA.
+ATTGCTGCAGACGCTCACCCCAGACACTCACTGCACCGGAGTGAGCGCGACCATCATGTCCATGCTCGTGGTCTTTCTCT
+TGCTGTGGGGTGTCACCTGGGGCCCAGTGACAGAAGCAGCCATATTTTATGAGACGCAGCCCAGCCTGTGGGCAGAGTCC
+GAATCACTGCTGAAACCCTTGGCCAATGTGACGCTGACGTGCCAGGCCCACCTGGAGACTCCAGACTTCCAGCTGTTCAA
+GAATGGGGTGGCCCAGGAGCCTGTGCACCTTGACTCACCTGCCATCAAGCACCAGTTCCTGCTGACGGGTGACACCCAGG
+GCCGCTACCGCTGCCGCTCGGGCTTGTCCACAGGATGGACCCAGCTGAGCAAGCTCCTGGAGCTGACAGGGCCAAAGTCC
+TTGCCTGCTCCCTGGCTCTCGATGGCGCCAGTGTCCTGGATCACCCCCGGCCTGAAAACAACAGCAGTGTGCCGAGGTGT
+GCTGCGGGGTGTGACTTTTCTGCTGAGGCGGGAGGGCGACCATGAGTTTCTGGAGGTGCCTGAGGCCCAGGAGGATGTGG
+AGGCCACCTTTCCAGTCCATCAGCCTGGCAACTACAGCTGCAGCTACCGGACCGATGGGGAAGGCGCCCTCTCTGAGCCC
+AGCGCTACTGTGACCATTGAGGAGCTCGCTGCACCACCACCGCCTGTGCTGATGCACCATGGAGAGTCCTCCCAGGTCCT
+GCACCCTGGCAACAAGGTGACCCTCACCTGCGTGGCTCCCCTGAGTGGAGTGGACTTCCAGCTACGGCGCGGGGAGAAAG
+AGCTGCTGGTACCCAGGAGCAGCACCAGCCCAGATCGCATCTTCTTTCACCTGAACGCGGTGGCCCTGGGGGATGGAGGT
+CACTACACCTGCCGCTACCGGCTGCATGACAACCAAAACGGCTGGTCCGGGGACAGCGCGCCGGTCGAGCTGATTCTGAG
+CGATGAGACGCTGCCCGCGCCGGAGTTCTCCCCGGAGCCGGAGTCCGGCAGGGCCTTGCGGCTGCGGTGCCTGGCGCCCC
+TGGAGGGCGCGCGCTTCGCCCTGGTGCGCGAGGACAGGGGCGGGCGCCGCGTGCACCGTTTCCAGAGCCCCGCTGGGACC
+GAGGCGCTCTTCGAGCTGCACAACATTTCCGTGGCTGACTCCGCCAACTACAGCTGCGTCTACGTGGACCTGAAGCCGCC
+TTTCGGGGGCTCCGCGCCCAGCGAGCGCTTGGAGCTGCACGTGGACGGACCCCCTCCCAGGCCTCAGCTCCGGGCGACGT
+GGAGTGGGGCGGTCCTGGCGGGCCGAGATGCCGTCCTGCGCTGCGAGGGACCCATCCCCGACGTCACCTTCGAGCTGCTG
+CGCGAGGGCGAGACGAAGGCCGTGAAGACGGTCCGCACCCCCGGGGCCGCGGCGAACCTCGAGCTGATCTTCGTGGGGCC
+CCAGCACGCCGGCAACTACAGGTGCCGCTACCGCTCCTGGGTGCCCCACACCTTCGAATCGGAGCTCAGCGACCCTGTGG
+AGCTCCTGGTGGCAGAAAGCTGATGCAGCCGCGGGCCCAGGGTGCTGTTGGTGTCCTCAGAAGTGCCGGGGATTCTGGAC
+TGGCTCCCTCCCCTCCTGTTGCAGCACAAGGCCGGGGTCTCTGGGGGGCTGGAGAAGCCTCCCTCATTCCTCCCAGGAAT
+TAATAAATGTGAAGAGAGGCTCTGTTTAAAATGTCTTTGGACTCCCAGGGCTGAGTGGGCTGGGATCTCGTGTCCTCAAA
+GTGGATGGGTTCTGGGGTGGCTCCTGAGGTAGAGGAGTGGAGAACTGGCTCTTAAGAGCACAGTTTTCTTTTCTTTTTTC
+TTTTTTGAGACAGGGTCTCACACTGTCACCCAGGCTGGAGTGCAGTGTTGCAGTCCGGCTCACTGCAGTTTTGACCTCCC
+AGGCTCAAGCGATCCTCCTGCCTCAGCCTCCCAAGTAGCTGGGAGCCTGGGCATGCATCGCCACGTCTGGCTAATTATTA
+TTTTTTGTAGACAGGGTCTCACTGTGTTGCCCAAGCTTGTCTTGAACTCCTGGCCTTAAGTGATCCTCCCACCTCAGCCT
+CCTGAGTAGCTGGGACTACAGGCATGAGCCACCATGCCTGGCCAACTCACATTTTTCTTTCTATTATTTATTTTTTGTAG
+AGATGAGTCTCACTATGTTGCCCAGGCTGGTCTTGAACTCTTGGGCTCAAGTGATCCTCCCGCCTCAGCCTCCCAAAGTG
+TTGGGATTACAGGTGTGAGCCACAGCACCTGGCCAACCAACACTTCTCAGGGCCTCTTTCATCTGTGCTCTTCCAGGATG
+CTGCCTCTTACTCCCTGGGCACCTCGGCCTGGTCCCAGCAGGTATGGGCAGTTGCTTGAGGCTCCAGACATACTCACCTC
+TACCTCGACCACATCAACCCCATCACCAAGAGGAGGTTCAGGGAAGCTGCATTTTGTGGTCTTGTCCTCCCAGTCCAGGG
+TGGTAGTGCTGGGCTGTTCCCAGCCTCCCACAGCAGAGCTGGACGCTGTGGAAATGGCTGGATTCCTCTGTGTTCTTTCC
+CATTATCCCTCAGCTCTGAGTCCTCTTGTGCCATCTGACATCTGATGCCTTCCCAACCACTGCCGAGTTTCTGCCTGGAG
+CAGGGCCTCAAGGCCCTGGCACACAGAAGATGCGTATCAGTATTATCAACCAATAGTTGATGAATTGTGTTTTTCAACGA
+ATTTGCTAGTGATCTGGTTTACTGCCTTAGTAATATCTAGTTCCTAATATGCCTATGCCTTTTAATGTCTGCACAGTCTA
+TGATGATGTCACTTCTCTCAAGCCTGATGTGTCCTCTCTCTCTTTCTTGCTAGTGGTTTATCAATTTTTAAAACCTTTTC
+TTGTTTATTTGTTTTTGAGACAGAATTTTGCTCTTGTTGCCCAGGCTGGAGTGCAATGGCGTGATCTTGGCTCACAGCAA
+CCTCTACCTCCTGGGTTCAAGTGATTCTCGTGCCTCAGCCTCCTGAGTAGCTGGGCTGACAGGCACCCACCACCACACCC
+AGCTAATTTTTATGTTTTTAGTAGAGATGGGGTTTCACCTTGTTGGGTGGCCAGGCTGGTCTCCAACTCCTGAGCTCAAG
+TGATCCACCCGCCTCAGCCTCCCAAAGTGCTGGGATTACAGGGGTGAGCCACCACGCCCAGCCCCAGCTTAGTTTTTTAA
+AAAGTTTATTTTAGCCATTCTAATAGGTATGTAGACATATCTAATAGTGGTTTTCCTTGCGTTTCCTTAATGTCTGATGA
+TGTTAAGCATTTTTTCCCCAAGTGCTTAGTTGCCATCTATATAGCATCTTTGATGAAATGTCTGTTTATATATTTTGCAC
+ACTTTAAAATATTGGGTTGTTT
