+++ tests/data/reference_sample_ligation.fna
@@ -1,0 +1,41 @@
+>XR_007073267.1 PREDICTED: Homo sapiens WAS protein family homolog 2-like (LOC124906207), misc_RNA
+CCTGGTTTGCAGACAGCCACCTTCTCTTCGCATCCTCACTTGGTGGGAAAAGACAGAGGAGAGAAAGAGAGAGAGAGAAA
+AATGAGATAGAGAGAGAGAGAGAAATGAGATAGAGAGAAATGAGAGAGAGAGAGAGAGAGAGAGAGAGAGGGAGACACAG
+AGACAATGCTCTCTTTTCTTACCAGATCTATAATGAGGGCCCCACCCCATGACCTCATCTAACCGTAATTACCTACCAAA
+GGCCCATCTCCACATACCATCACATTAGGGGTTAGGGTTTCAACATAAGCATTTGGAGAGGACATAAACATTCTGTCTAC
+AACATGAGTGGAGATCCATCTCTTCTTTACCTCTGGTAAGGGGACCACACGCTGCAGCCAGCGAGACAGTGGCATGTTCT
+TGTTACAACTCGATCTAACTCCCCCAGAAGAGGAGGCAGGGAAGGCGGACAAAACTGGGAGAGGGAGAGAGTGTTAGGAA
+GAGAGTAGGGTGGCCAGAGGCAGCAAATAAAATATAAAATGCTTAATTCTGAATCTCAGATAAACAACCAATAATGTTTT
+TTAGCATAAGTATGTCCCAAACTAAGCTTGGGACATATTTACGCTACGAAATTATTCGTTGTTTATCTAAAATTCAAACT
+AGCTGGGCATCCTGTCTTTTAATCTGGCAACCCTAAAAGGCAAGGGCCAAAAATGCCGGAGGCAAGCCAACGGATTCCAG
+GAGGGACAACTGCTGGACTTGACTGATGATGCTCTTTTTATATATTTAACTTTTTAAAAAAGCCTCTTTTCTTTCTTTTT
+ACCAGCTTTTCACTAGCTTTTTAAAAATTGTGGTAAAACATACATAACCTAAAATTCACCGTCTAAACCATTTTTCAGCA
+TACAGTTCAGTGGGATTAAGTAGATTCACACCGTTGTGCCGCCATCACTACCACTCATTTCCAGCACCCTTCCATCACCC
+CAGCCTGAAACTCTACCCATTGAACATGAGCTGCCCAACTCCGCCTTGCTTCCCCATTCCTGGCGACCACTGCTTCTGTC
+TCTGTGAATTTTGACTATTCTAGGCACTTCACAAAACTGGACTCATACGATATCTGTAGTTTTGCGTCTGGCTTCTCTAT
+TTAATTCTTAAAGGGGGGTGGGAACTAAGCAGATCACAAGGGAGCTGCCCACAGAGGTAAAGACAAGGTCAGGTAGGCTG
+AGAGACGCAGGAAAGCGGGTCAAGGCGTAGGGCTGGAGGGCAGGGGCGGGCCCTGGGCCTGGGCTGGGGGTCCTGCCCCG
+GGGCGCACCCCGGGCAAGGGCTGCCCGGAGGAGCCGAGGCTGGCGGACAGCTTGGCCCTGAGCTTGAGGGGAAGGCAGCG
+ATGGGACAAAGGACGGAGGTCTGGGAAGAGGGTCTGCAGAGCAGAAAGCACGGGTAGGGGCGGCCTGACGCTCGGAAGAC
+AACGCATGGGAGCCGTGTGCACGTCGGGAGCTCGGAGTGAGCGCACCATGACTCCTGTGAGGATGCAGCACTCCCTGGCA
+GGTCAGACCTATGCCGTGCCCCTCATCCAGCCAGACCTGCGGCGAGAGGAGGCCGTCCAGCAGATGGCGGATGCCCTGCA
+GTACCTGCAGAAGGTCTCTGGAGACATCTTCAGCAGGATCTCCCAGCAGGTAGAGCAGAGCCGGAGCCAGGTGCAGGCCA
+TTGGAGAGAAGGTCTCCTTGGCCCAGGCCAAGATTGAGAAGATCAAGGGCAGCAAGAAGGCCATCAAGGTGTTCTCCAGT
+GCCAAGTACCCTGCTCCAGAGCACCTGCAGGAATATGGCTCCATCTTCACGGGCGCCCAGGACCCTGGCCTGCAGAGACG
+CTCCCGCCACAGGATCCAGAGCAAGCACCGCCCCCTGGATGAGCGGGCCCTGCAGGAGAAGCTGAAGGAATTTCCTGTGT
+GCGTGAGCACCAAGCCGGAGCCTGAGGACGATGCAGAAGAGGGACTTGGGGGTCTTCCCAGCAACATCAGCTCTGTCAGC
+TCCTTGCTGCTTTTCAACACCACCGAGAACCTGTAGAAGAAGTATGTCTTCCTGGACCCCCTGGCTGGTGCTGTAACAAA
+GACCCATGTGATGCTGGGGGCAGAGACAGAGGAGAAGCTGTTTGATGCCCCCTTGTCCATCAGCAAGAGAGAGCAGCTGG
+AACAGCAGGTCCCAGAGAACTACTTCTATGTGCCAGACCTGGGCCAGGTGCCTGAGATTGATGTTCCATCCTACCTGCCT
+GACCTGCCCGGCATTGCCAACGACCTCATGTACATTGCCGACCTGGGCCCCGGCATTGCCCCCTCTGCCCCTGGCACCAT
+TCCAGAACTGCCCACCTTCCACACTGAGGTAGCCGAGCCTCTCAAGGCAGACCTACAAGATGGGGTACTAACACCACCCC
+CACCGCCCCCACCACCACCCCCAGCTCCTGAGGTGCTGGCCAGTGCACCCCCACTCCCACCCTCAACCGCGGCCCCTGTA
+GGCCAAGGCGCCAGGCAGGACGACAGCAGCAGCAGCACGTCTCCTTCAGTCCAGGGAGCTCCCAGGGAAGTGGTCGACCC
+CTCCGGTGGCTGGGCCACTCTGCTAGAGTCCATCCGCCAAGCTGGGGGCATCGGCAAGGCCAAGCTGCGCAGCATGAAGG
+AGCGAAAGCTGGAGAAGAAGAAGCAGAAGGAGCAGGAGCAAGTGAGAGCCACGAGCCAAGGTGGGCACTTGATGTCGGAT
+CTCTTCAACAAGCTGGTCATGAGGCACAAGGGCATCTCTGGGAAAGGACCTGGGGCTGGTGAGGGGCCCGGAGGAGCCTT
+TGCCCGCGTGTCAGACTCCATCCCTCCTCTGCCGCCACCGCAGCAGCCACAGGTAGATGAGGACAAGGACGACTGGGAAT
+CCTAGGGGGCTCCATGACACCTTCCCCCCCAGACCCAGACTTGGGCCGTTGCTCTGACATGGACACAGCCAGGACAAGCT
+GCTCAGACCTGCTTCCCTGGGAGGGGGTGACGGAACCAGCACTGTGTGGAGACCAGCTTCAAGGAGCGGAAGGCTGGCTT
+GAGGCCACACAGCTGGGGCGGGGACTTCTGTCTGCCTGTGCTCCATGGGGGGACGGCTCCACCCAGCCTGCGCCACTGTG
+TTCTTAAGAGGCTTCCAGAGAAAACGGCACACCAATCAATAAAGAACTGAGCAGAAA
